[fix] fixed error at mask colors
