+++ frontend/app/src/main/java/com/project/spire/ui/create/CanvasViewModel.kt
@@ -12,15 +12,6 @@ import com.project.spire.utils.PaintOptions
 
 class CanvasViewModel: ViewModel() {
 
-    /*
-    TODO
-    구현한 것들 테스트 안해봄
-    WriteTextActivity -> Fragment로 변경, ImageEditActivity와 ViewModel을 공유하도록
-    Button Click 시 UtilBase64 활용하여 Image / Mask -> Base64 String으로 변환
-    Inference Server에 전송
-    ImageView를 만들어 Response 출력
-     */
-
     private val _originImageBitmap = MutableLiveData<Bitmap>()
     val originImageBitmap: LiveData<Bitmap>
         get() = _originImageBitmap
@@ -29,8 +20,8 @@ class CanvasViewModel: ViewModel() {
         _originImageBitmap.value = bitmap
     }
 
-    private val STROKE_PEN = 50f
-    private val STROKE_ERASER = 70f
+    private val STROKE_PEN = 60f
+    private val STROKE_ERASER = 80f
     private val MODE_CLEAR = PorterDuffXfermode(PorterDuff.Mode.CLEAR) // clears when overlapped
     private val ALPHA = 150
